move invite update views to accounts app and minor fixes
