+++ accounts/templates/profile/detail.html
@@ -73,7 +73,7 @@
         </div>
         <div class="lists-menu-buttons-container">
             {% if current_user.id == profile_user.id %}
-                <form class="lists-menu-buttons-form" method="POST" action="{% url 'appsite:invite_update_all' profile_user.id %}">
+                <form class="lists-menu-buttons-form" method="POST" action="{% url 'accounts:invite_update_all' profile_user.id %}">
                     {%csrf_token%}
                     <input type="submit" class="accept lists-menu-button" name="accept" value="Aceitar tudo" />
                     <input type="submit" class="refuse lists-menu-button" name="refuse" value="Recusar tudo" />
