+++ appsite/urls.py
@@ -6,8 +6,7 @@ app_name='appsite'
 urlpatterns = [
     ### Profile ###
 
-    path('user/<int:pk>/invite/',      views.InviteUpdateView.as_view(),    name='invite_update'),
-    path('user/<int:pk>/invite/all/',  views.InviteUpdateAllView.as_view(), name='invite_update_all'),
+
     path('user/<int:pk>/list/create/', views.ListCreateView.as_view(),      name='list_create'), 
 
     ### List ###
